requirements.txt
streamlit==1.32.0
google-generativeai>=0.4.0
spacy==3.7.4
nltk==3.8.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.1
plotly==5.19.0
